requirements.txt
pi==0.1.2
Pillow==9.4.0
